colors updated

+++ app/src/main/java/com/firstapp/go2meet_map/Item.java
@@ -95,11 +95,11 @@ public class Item {
         return isFree;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
